Add a constructor for babel-jest transforms
