+++ packages/craco/lib/features/test/jest-babel-transform.js
@@ -1,38 +1,6 @@
-const babelJest = require("babel-jest");
-
 const { loadCracoConfig } = require("../../config");
-const { isArray } = require("../../utils");
-
-const craBabelTransformer = {
-    presets: ["babel-preset-react-app"],
-    babelrc: false,
-    configFile: false
-};
-
-const context = {
-    env: process.env.NODE_ENV
-};
+const { createJestBabelTransform } = require("./create-jest-babel-transform");
 
 const cracoConfig = loadCracoConfig(context);
 
-const { addPresets, addPlugins } = cracoConfig.jest.babel;
-
-if (cracoConfig.babel) {
-    if (addPresets) {
-        const { presets } = cracoConfig.babel;
-
-        if (isArray(presets)) {
-            craBabelTransformer.presets = craBabelTransformer.presets.concat(presets);
-        }
-    }
-
-    if (addPlugins) {
-        const { plugins } = cracoConfig.babel;
-
-        if (isArray(plugins)) {
-            craBabelTransformer.plugins = plugins;
-        }
-    }
-}
-
-module.exports = babelJest.createTransformer(craBabelTransformer);
+module.exports = createJestBabelTransform(cracoConfig);
